+++ packages/components/src/Sales/Direct/Form.tsx
@@ -26,7 +26,7 @@ import {
   CreateOfferStep,
   formatDateDatetime,
   formatError,
-  parsePrice,
+  parseBigNumber,
   useCreateOffer,
 } from '@nft/hooks'
 import { FaInfoCircle } from '@react-icons/all-files/fa/FaInfoCircle'
@@ -117,7 +117,8 @@ const SalesDirectForm: VFC<Props> = ({
     if (!c) throw new Error("Can't find currency")
     return c
   }, [currencies, currencyId])
-  const priceUnit = parsePrice(price, currency.decimals)
+  const priceUnit = parseBigNumber(price, currency.decimals)
+  const quantityBN = parseBigNumber(quantity)
 
   const amountFees = useMemo(() => {
     if (!price) return BigNumber.from(0)
@@ -144,7 +145,7 @@ const SalesDirectForm: VFC<Props> = ({
       onOpen()
       const id = await createAndPublishOffer({
         type: 'SALE',
-        quantity: BigNumber.from(quantity),
+        quantity: quantityBN,
         unitPrice: priceUnit,
         assetId: asset.id,
         currencyId: currency.id,
@@ -198,20 +199,30 @@ const SalesDirectForm: VFC<Props> = ({
             clampValueOnBlur={false}
             min={0}
             step={Math.pow(10, -currency.decimals)}
-            precision={currency.decimals}
             allowMouseWheel
             w="full"
             onChange={(x) => setValue('price', x)}
-            format={(e) => e.toString()}
           >
             <NumberInputField
               id="price"
               placeholder={t('sales.direct.form.price.placeholder')}
               {...register('price', {
                 required: t('sales.direct.form.validation.required'),
-                validate: (value) =>
-                  parseFloat(value) > 0 ||
-                  t('sales.direct.form.validation.positive'),
+                validate: (value) => {
+                  const splitValue = value.split('.')
+
+                  if (parseFloat(value) <= 0) {
+                    return t('sales.direct.form.validation.positive')
+                  }
+                  if (
+                    splitValue[1] &&
+                    splitValue[1].length > currency.decimals
+                  ) {
+                    return t('sales.direct.form.validation.decimals', {
+                      nbDecimals: currency.decimals,
+                    })
+                  }
+                },
               })}
             />
             <NumberInputStepper>
@@ -252,17 +263,28 @@ const SalesDirectForm: VFC<Props> = ({
               clampValueOnBlur={false}
               min={1}
               max={quantityAvailable?.toNumber()}
-              step={1}
               allowMouseWheel
               w="full"
               onChange={(x) => setValue('quantity', x)}
-              format={(e) => e.toString()}
             >
               <NumberInputField
                 id="quantity"
                 placeholder={t('sales.direct.form.quantity.placeholder')}
                 {...register('quantity', {
                   required: t('sales.direct.form.validation.required'),
+                  validate: (value) => {
+                    if (
+                      parseFloat(value) < 1 ||
+                      parseFloat(value) > quantityAvailable?.toNumber()
+                    ) {
+                      return t('sales.direct.form.validation.in-range', {
+                        max: quantityAvailable?.toNumber(),
+                      })
+                    }
+                    if (!/^\d+$/.test(value)) {
+                      return t('sales.direct.form.validation.integer')
+                    }
+                  },
                 })}
               />
               <NumberInputStepper>
@@ -394,7 +416,7 @@ const SalesDirectForm: VFC<Props> = ({
                 color="brand.black"
                 mx={1}
                 fontWeight="semibold"
-                amount={quantity ? amountFees.mul(quantity) : 0}
+                amount={amountFees.mul(quantityBN)}
                 currency={currency}
               />
             </Text>
@@ -408,7 +430,7 @@ const SalesDirectForm: VFC<Props> = ({
                 color="brand.black"
                 ml={1}
                 fontWeight="semibold"
-                amount={quantity ? amountRoyalties.mul(quantity) : 0}
+                amount={amountRoyalties.mul(quantityBN)}
                 currency={currency}
               />
             </Text>
@@ -420,7 +442,7 @@ const SalesDirectForm: VFC<Props> = ({
                 color="brand.black"
                 mx={1}
                 fontWeight="semibold"
-                amount={quantity ? priceWithFees.mul(quantity) : 0}
+                amount={priceWithFees.mul(quantityBN)}
                 currency={currency}
               />
             </Text>