+++ packages/hooks/src/index.ts
@@ -28,6 +28,7 @@ import { formatAddress } from './utils/formatAddress'
 import { formatError } from './utils/formatError'
 import { formatSocial } from './utils/formatSocial'
 import { parsePrice } from './utils/parsePrice'
+import { parseBigNumber } from './utils/parseBigNumber'
 import { removeEmptyFromObject } from './utils/removeEmptyFromObject'
 
 export { LiteflowProvider } from './context'
@@ -91,6 +92,7 @@ export {
   isSameAddress,
   removeEmptyFromObject,
   parsePrice,
+  parseBigNumber,
   formatError,
 }
 export type { ISessionContext }