+++ packages/templates/src/home/Explorer.tsx
@@ -5,6 +5,7 @@ import {
   Checkbox,
   Flex,
   FormControl,
+  FormErrorMessage,
   FormLabel,
   Grid,
   GridItem,
@@ -28,7 +29,7 @@ import {
   TokenCard,
   wrapServerSideProps,
 } from '@nft/components'
-import { parsePrice, removeEmptyFromObject } from '@nft/hooks'
+import { parseBigNumber, removeEmptyFromObject } from '@nft/hooks'
 import { GetServerSideProps } from 'next'
 import Trans from 'next-translate/Trans'
 import useTranslation from 'next-translate/useTranslation'
@@ -145,7 +146,7 @@ const minPriceFilter = (
         availableQuantity: { greaterThan: '0' },
         currencyId: { equalTo: currency.id },
         unitPrice: {
-          greaterThanOrEqualTo: parsePrice(
+          greaterThanOrEqualTo: parseBigNumber(
             minPrice.toString(),
             currency.decimals,
           ).toString(),
@@ -166,7 +167,7 @@ const maxPriceFilter = (
         availableQuantity: { greaterThan: '0' },
         currencyId: { equalTo: currency.id },
         unitPrice: {
-          lessThanOrEqualTo: parsePrice(
+          lessThanOrEqualTo: parseBigNumber(
             maxPrice.toString(),
             currency.decimals,
           ).toString(),
@@ -505,51 +506,106 @@ export const Template: VFC<
 
               {currency && (
                 <Flex gap={3}>
-                  <InputGroup>
-                    <NumberInput
-                      clampValueOnBlur={false}
-                      min={0}
-                      step={Math.pow(10, -currency.decimals)}
-                      precision={currency.decimals}
-                      allowMouseWheel
-                      w="full"
-                      isDisabled={isSubmitting}
-                      onChange={(x: any) => setValue('minPrice', x)}
-                      format={(e) => e.toString()}
-                    >
-                      <NumberInputField
-                        placeholder={t('explore.form.min-price.placeholder')}
-                        {...register('minPrice')}
-                      />
-                      <NumberInputStepper>
-                        <NumberIncrementStepper />
-                        <NumberDecrementStepper />
-                      </NumberInputStepper>
-                    </NumberInput>
-                  </InputGroup>
+                  <FormControl isInvalid={!!errors.minPrice}>
+                    <InputGroup>
+                      <NumberInput
+                        clampValueOnBlur={false}
+                        min={0}
+                        step={Math.pow(10, -currency.decimals)}
+                        allowMouseWheel
+                        w="full"
+                        isDisabled={isSubmitting}
+                        onChange={(x: any) => setValue('minPrice', x)}
+                      >
+                        <NumberInputField
+                          id="minPrice"
+                          placeholder={t('explore.form.min-price.placeholder')}
+                          {...register('minPrice', {
+                            validate: (value) => {
+                              if (!value) return
+                              const splitValue = value.toString().split('.')
 
-                  <InputGroup>
-                    <NumberInput
-                      clampValueOnBlur={false}
-                      min={0}
-                      step={Math.pow(10, -currency.decimals)}
-                      precision={currency.decimals}
-                      allowMouseWheel
-                      w="full"
-                      isDisabled={isSubmitting}
-                      onChange={(x: any) => setValue('maxPrice', x)}
-                      format={(e) => e.toString()}
-                    >
-                      <NumberInputField
-                        placeholder={t('explore.form.max-price.placeholder')}
-                        {...register('maxPrice')}
-                      />
-                      <NumberInputStepper>
-                        <NumberIncrementStepper />
-                        <NumberDecrementStepper />
-                      </NumberInputStepper>
-                    </NumberInput>
-                  </InputGroup>
+                              if (value < 0) {
+                                return t(
+                                  'explore.form.min-price.validation.positive',
+                                )
+                              }
+                              if (
+                                splitValue[1] &&
+                                splitValue[1].length > currency.decimals
+                              ) {
+                                return t(
+                                  'explore.form.min-price.validation.decimals',
+                                  {
+                                    nbDecimals: currency.decimals,
+                                  },
+                                )
+                              }
+                            },
+                          })}
+                        />
+                        <NumberInputStepper>
+                          <NumberIncrementStepper />
+                          <NumberDecrementStepper />
+                        </NumberInputStepper>
+                      </NumberInput>
+                    </InputGroup>
+                    {errors.minPrice && (
+                      <FormErrorMessage>
+                        {errors.minPrice.message}
+                      </FormErrorMessage>
+                    )}
+                  </FormControl>
+                  <FormControl isInvalid={!!errors.maxPrice}>
+                    <InputGroup>
+                      <NumberInput
+                        clampValueOnBlur={false}
+                        min={0}
+                        step={Math.pow(10, -currency.decimals)}
+                        allowMouseWheel
+                        w="full"
+                        isDisabled={isSubmitting}
+                        onChange={(x: any) => setValue('maxPrice', x)}
+                      >
+                        <NumberInputField
+                          id="maxPrice"
+                          placeholder={t('explore.form.max-price.placeholder')}
+                          {...register('maxPrice', {
+                            validate: (value) => {
+                              if (value === null) return
+                              const splitValue = value.toString().split('.')
+
+                              if (value < 0) {
+                                return t(
+                                  'explore.form.max-price.validation.positive',
+                                )
+                              }
+                              if (
+                                splitValue[1] &&
+                                splitValue[1].length > currency.decimals
+                              ) {
+                                return t(
+                                  'explore.form.max-price.validation.decimals',
+                                  {
+                                    nbDecimals: currency.decimals,
+                                  },
+                                )
+                              }
+                            },
+                          })}
+                        />
+                        <NumberInputStepper>
+                          <NumberIncrementStepper />
+                          <NumberDecrementStepper />
+                        </NumberInputStepper>
+                      </NumberInput>
+                    </InputGroup>
+                    {errors.maxPrice && (
+                      <FormErrorMessage>
+                        {errors.maxPrice.message}
+                      </FormErrorMessage>
+                    )}
+                  </FormControl>
                 </Flex>
               )}
             </Stack>